cli/agents/industry_analysis_agent/pyproject.toml
[project]
name = "meeting-industry-analysis-agent"
version = "0.1.0"
description = "Industry Analysis Agent for Meeting Preparation - analyzes industry trends and opportunities"
requires-python = ">=3.12"
dependencies = [
    "genai-protocol>=1.0.9",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
